requirements.txt
asgiref==3.5.2
Django==4.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework
